+++ src/main/java/io/github/scru128/Scru128Generator.java
@@ -167,7 +167,9 @@ public class Scru128Generator implements Iterable<@NotNull Scru128Id>, Iterator<
      * generation method and this method to avoid race conditions.
      *
      * @return status code from the last generation of ID.
+     * @deprecated Use {@link #generateMonotonic()} to guarantee monotonicity.
      */
+    @Deprecated
     public @NotNull Status getLastStatus() {
         return lastStatus;
     }
@@ -206,7 +208,10 @@ public class Scru128Generator implements Iterable<@NotNull Scru128Id>, Iterator<
 
     /**
      * Status code returned by {@link #getLastStatus()} method.
+     *
+     * @deprecated Use {@link #generateMonotonic()} to guarantee monotonicity.
      */
+    @Deprecated
     public enum Status {
         /**
          * Indicates that the generator has yet to generate an ID.