+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - `generateNoRewind()` and `generateCoreNoRewind(long, long)` to
   `Scru128Generator` (experimental)
 
+### Maintenance
+
+- Updated gradlew and compile-time/development dependencies
+
 ## v2.2.0 - 2022-12-23
 
 ### Added
@@ -16,7 +20,7 @@
 
 ### Maintenance
 
-- Update gradlew and compile-time/development dependencies
+- Updated gradlew and compile-time/development dependencies
 
 ## v2.1.2 - 2022-06-11