+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v2.2.1 - unreleased
+## v2.2.1 - 2023-03-19
 
 ### Added
 
@@ -10,6 +10,7 @@
 ### Maintenance
 
 - Updated gradlew and compile-time/development dependencies
+- Improved documentation about generator method flavors
 
 ## v2.2.0 - 2022-12-23